+++ Source/Scene/LabelCollection.js
@@ -314,7 +314,6 @@ define([
         this._textureAtlas = undefined;
 
         this._billboardCollection = new BillboardCollection();
-        this._billboardCollection.clampToPixel = true;
         this._billboardCollection.setDestroyTextureAtlas(false);
 
         this._spareBillboards = [];
@@ -359,6 +358,15 @@ define([
          * });
          */
         this.modelMatrix = Matrix4.IDENTITY.clone();
+
+        /**
+         * If true, aligns all vertices to a pixel in screen space,
+         * providing crisper text at the cost of jumpier motion.
+         * Defaults to true.
+         *
+         * @type Boolean
+         */
+        this.clampToPixel = true;
 
         /**
          * The current morph transition time between 2D/Columbus View and 3D,
@@ -594,6 +602,7 @@ define([
 
         billboardCollection.modelMatrix = this.modelMatrix;
         billboardCollection.morphTime = this.morphTime;
+        billboardCollection.clampToPixel = this.clampToPixel;
 
         var rebindAllGlyphsInAllLabels = false;
         if (++this._frameCount % 100 === 0) {